backend/requirements.txt
django==5.2.9
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
psycopg2-binary==2.9.11
dj-database-url==3.0.1
python-dotenv==1.2.1
drf-spectacular==0.29.0
django-cors-headers==4.9.0
pytest==9.0.2
pytest-django==4.11.1
